scripts/pyproject.toml
[tool.poetry]
name = "scripts"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
PyGithub = "^2.3.0"
ipykernel = "^6.29.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
